fix(security): remediate 114 code scanning alerts — Netty/Logback/Jackson upgrades + CodeQL fixes (#156)
**Dependency upgrades (removes ~80 OWASP/Trivy alerts):**
- Netty: 4.1.119.Final → 4.1.131.Final
  Fixes CVE-2025-55163 (fixed in 4.1.124), CVE-2025-58056, CVE-2025-58057,
  CVE-2025-67735 (fixed in 4.1.129) across api, license, sync services
- Logback: 1.5.18 → 1.5.32 in all 3 backend services
- Jackson-core: 2.19.0 → 2.19.4 (latest 2.19.x; GHSA-72hv pending fix in 2.21.1)
- Add Netty force resolution to license service (was missing)
- Add Jackson force resolution to sync service (was missing)

**CodeQL code fixes (resolves 5 warning-level alerts):**
- AdminViewModel.kt: fix integer-multiplication-cast-to-long — olderThanDays.toLong()
  prevents theoretical Int overflow for very large day values before Long cast
- NavigationItems.kt: remove constant-comparison — groupCount is always > 0
  when currentGroup != null (groupCount is set to 1 on every new group start)
- ZyntaStatCard.kt: add explicit parentheses to clarify operator precedence
  in chart Y coordinate calculation (whitespace-contradicts-precedence)
- DesktopA4PrintDelegate.kt: use Files.createTempFile (NIO) with rw------- POSIX
  permissions — fixes local-temp-file-or-directory-information-disclosure
- CloudPrntClient.kt: replace deprecated URL(String) constructor with
  URI.create().toURL() — fixes java/deprecated-call (Java 20+ deprecation)

**GitHub Security tab dismissals (74 false positives):**
- java/field-masks-super-field (13): Kotlin sealed class data objects do not
  actually mask parent fields — CodeQL Java analyzer false positive
- java/local-variable-is-never-read (59): Kotlin MVI Intent/Effect sealed class
  data objects flagged incorrectly by Java analyzer
- java/random-used-once (1): SecureRandom().nextBytes(iv) is correct for crypto IV
- java/unused-reference-type (1): Koin DSL extension usage invisible to Java analyzer

+++ composeApp/feature/admin/src/commonMain/kotlin/com/zyntasolutions/zyntapos/feature/admin/AdminViewModel.kt
@@ -158,7 +158,7 @@ class AdminViewModel(
 
     private suspend fun purgeExpiredData(olderThanDays: Int) {
         updateState { copy(isLoading = true) }
-        val cutoff = Clock.System.now().toEpochMilliseconds() - olderThanDays * 24 * 60 * 60 * 1000L
+        val cutoff = Clock.System.now().toEpochMilliseconds() - olderThanDays.toLong() * 24 * 60 * 60 * 1000L
         when (val result = systemRepository.purgeExpiredData(cutoff)) {
             is Result.Success -> {
                 updateState {
